refactor: toolbar component
