+++ .dumi/theme/pages/Example/index.module.less
@@ -12,6 +12,8 @@
 
   .container {
     flex: 1;
+    background-color: #fff;
+
     .sider {
       height: 100%;
 
@@ -29,6 +31,7 @@
 
         .headerContainer {
           height: 100%;
+          background-color: #fff;
 
           .title {
             font-weight: 500;
@@ -38,13 +41,10 @@
 
       .preview {
         height: 100%;
-        border: solid 2px red;
-
       }
 
       .editor {
         height: 100%;
-        border: solid 2px green;
       }
     }
