+++ .dumi/theme/slots/CodeEditor/index.module.less
@@ -1,0 +1,3 @@
+.editor {
+  
+}
