+++ .dumi/theme/slots/ExamplePreview/index.module.less
@@ -1,0 +1,24 @@
+.preview {
+  width: 100%;
+  height: 100%;
+  position: relative;
+
+  padding: 16px;
+  overflow: auto;
+
+  .container {
+    width: 100%;
+    height: 100%;
+    position: relative;
+  }
+  // 盖在 DEMO 运行的 container 上
+  .result {
+    width: 100%;
+    height: 100%;
+    position: absolute;
+    top: 0;
+    left: 0;
+
+    background: #fff;
+  }
+}
