+++ .dumi/theme/slots/PlayGround/index.module.less
@@ -4,12 +4,15 @@
 .playground {
   margin: 0;
   padding: 0;
-  border: 1px solid @border-color-split;
+  // border: 1px solid @border-color-split;
   border-radius: 4px;
   background: #fff;
   height: @editor-height;
   overflow: hidden;
   position: relative;
+
+  width: 100%;
+  height: 100%;
 
   .preview,
   .editor {
@@ -42,8 +45,55 @@
       overflow: auto;
     }
 
-    .Resizer.vertical {
-      width: 11.5px !important;
+    .Resizer {
+      background: #000;
+      opacity: 0.1;
+      z-index: 1;
+      box-sizing: border-box;
+      background-clip: padding-box;
+    
+      &:hover {
+        transition: all 2s ease;
+      }
+    
+      &.horizontal {
+        height: 11px;
+        margin: -5px 0;
+        border-top: 5px solid rgba(255, 255, 255, 0);
+        border-bottom: 5px solid rgba(255, 255, 255, 0);
+        cursor: row-resize;
+        width: 100%;
+    
+        &:hover {
+          border-top: 5px solid rgba(0, 0, 0, 0.5);
+          border-bottom: 5px solid rgba(0, 0, 0, 0.5);
+        }
+      }
+    
+      &.vertical {
+        width: 11px;
+        margin: 0 -5px;
+        border-left: 5px solid rgba(255, 255, 255, 0);
+        border-right: 5px solid rgba(255, 255, 255, 0);
+        cursor: col-resize;
+    
+        Pane2 {
+          border-left: 1px solid #e6e6e6;
+        }
+    
+        &:hover {
+          border-left: 5px solid rgba(0, 0, 0, 0.5);
+          border-right: 5px solid rgba(0, 0, 0, 0.5);
+        }
+      }
+    
+      &.disabled {
+        cursor: not-allowed;
+    
+        &:hover {
+          border-color: transparent;
+        }
+      }
     }
   }
